requirements.txt
asgiref==3.7.2
Django==4.2.4
django-js-asset==2.1.0
django-mptt==0.14.0
djangorestframework==3.14.0
exceptiongroup==1.1.2
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
